Flesh out a usable implementation of RealmConnector that stores its database information in an XML file.
org.apache.tomcat.security.file.FileRealmTool is a command line
application that can be used to manipulate the contexts of an XML
database, in preparation for use by Tomcat.  For example, you might
add a new user, and put them in a specified group, like this:

	java org.apache.tomcat.security.file.FileRealmTool \
	  mydatabase.xml \
	  -addUser hans mypassword \
	  -addMember committer hans \
	  -addMember pmc hans

See the JavaDoc comments on FileRealmTool for all the possible commands.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281092 13f79535-47bb-0310-9956-ffa450edef68