Changed the names of the attributes to reflect the original command line ones.
WARNING: this might break compatibility so you want to check
the new attributes names if you are using the Javadoc2 task.
(jon, might be your case).

Sorry for this, but I think it's better to have the same names on the
command line and on the XML file, even if this breaks java
naming conventions.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@264293 13f79535-47bb-0310-9956-ffa450edef68