Tomcat is v3.0 now.
1.0 is not appropriate because we will be going back from 2.2 which was in
the "development" tree.

We don't want Tomcat tied to the version numbers of any of the specifications
it implements - that rules out 1.1 and 2.2.

3.0 is the last stable release for 1999. Bug fixes updates to this release
will be of the form 3.0.[z]. Feature updates to this release will be of the
form 3.[y]. If we implement a newer version of the JSP or Servlet specs, then
we'll rev the major version number.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281291 13f79535-47bb-0310-9956-ffa450edef68