Infrastructure and error recovery in preparation for scripting support
With this change, <jsp:scriptlet > is accepted (note the white space) and
<jsp scriptlet runat="server"> produces an error.

Also, every ParseEventListener method is now passed a Hashtable of
attributes with the exception of handleComment.  This facilitates
rapid response to JSP spec changes, and will enable future tag libraries
to reuse more of the current Jasper implementation.

No regressions were noted in watchdog jsp tests as a result of this change.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281315 13f79535-47bb-0310-9956-ffa450edef68