- Moved code around in Context - grouped all class loading related stuff. - Moved ServletClassLoader to an interface and ServletClassLoaderImpl with the original code.
IMPORTANT:
- I have no ideea why we have 2 different classloaders in Context -
it seems the "normal" class loader is used only once in JSP, and never set.

- Both methods in ServletClassLoader can be replaced with "standard" code -
classpath is in Context ( and you get the loader calling a method in context,
so there is no need - IMHO), and loadServlet is just calling loadClass.
Sam, Anil - please take a look.

- ClassLoader implementation should go in a different package ( tomcat.support?)

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281346 13f79535-47bb-0310-9956-ffa450edef68