Moved logic from Context into ContextInterceptors ( no interface yet, still hardcoded until initialization cleanup).
Context interceptors are grouped in tomcat.context ( if you have a better name
please move them !).

Still a lot of work to clean the interceptors, or start from zero with new
ones, but Context is much better now.

Since Context has a lot of properties ( as required by Specs ), it is better
if Context will have little logic inside, but delegate to specialized
classes.

Also, the new methods added to Context will simplify an admin interface.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281349 13f79535-47bb-0310-9956-ffa450edef68