Removed old HttpAdapter ( unused ). HttpServer will turn into an interface that will work with any adapter and has no more "standalone" code in it.
Also, removed the old "adapter" hack ( load a properties file from
classpath ) - XML can set it up now.

The code still exists - the new HttpAdapter ( in tomcat.service.http)
is based on the code in tomcat.server ( with cosmetic changes and following
the Adapter interfaces ). The code here was the original, in case something
gets wrong.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281351 13f79535-47bb-0310-9956-ffa450edef68