One step towards Craig's design for sessions.
The session manager no longer needs to process the request, it
acts just as a repository.
The request processing ( extracting the session Id) is done by
an interceptor. The other way - setting the response after a new
session is created should also be in a post-service interceptor.
( this way you can configure SSL sessions or URL rewriting without
touching DbManager )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281362 13f79535-47bb-0310-9956-ffa450edef68