Cleaned-up RequestDispatcher !
I don't know if the previous model was better or not, but it was _very_ hard to
follow it.

When you call getRequestDispatcher a new Request ( used to be LookupResult, but it behaved the
same way ) is created. Now this request is used when invoking the included servlet,
instead of setting the old request with some of the parameters in the new request.

While the old way was probably faster, it required a number of hacks ( the included
request had special "magic" in it in form of special attributes).

Feel free to find a better way to implement include().

This will also clean up most of InvokerServlet ( which can move to a normal interceptor)

It seems all test are passing ( even with the incomplete implementation we have - that means
we need more tests, since the code is obviously broken !)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281384 13f79535-47bb-0310-9956-ffa450edef68