Request and Response are now interfaces.
Note that it is an "evolution" - i.e. the methods in the current interfaces
are based on what it is used in the current tomcat, not in what it should
be.

Please start reviewing the interfaces, and add documentation to the methods
that are right and remove the methods that are wrong ( fixing tomcat to use
the good methods ).

Also, added ContextInterceptor that will handle context initialization and
destruction.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281389 13f79535-47bb-0310-9956-ffa450edef68