Important change in XML reading ( for configuration ).
The main ideea is to decouple the XML architecture from the Java objects.
XML documents should be easy to read/edit, Java object have different
requirement, and IMHO it's a bad ideea to design a Java object as a document
or a document as a programming language.

The XML object will be read ( by a SAX driver ) and various informations will
be extracted - creating Java objects or setting properties ( like in Ant).

The model is the same - except that we have to deal with object hierarchies,
and we want to reduce the amount of "magic". The reader will have no
"behind-the-scene" role, all object creations, property settings, or relations
must be explicitly requested. ( either as Java code or using <taskdef-like>)

There are few "pre-defined" actions:
- create object
- set parent
- add child
- set object properties from tags
- invoke an object method ( like addMapping) using body or sub-elements as
parameters
- set properties using the body ( usefull in web.xml-like documents )

You can define new actions, using the SaxContext.

This is different from XSL or XML serialization - we are just extracting
some informations from an XML file ( that is not under our control - so
serialization is not posible ) - we don't process the information in any
way except feeding it to Java objects.

That means all validation is done by the Java objects ( vs. DTD ), we
don't need the full document in memory - Java object hierarchy instead
of DOM, and all processing is inside java code instead of templates.

BTW, with all watchdog tests deployed, the time to read all web.xml files
is 3 sec versus 6 sec. with the DOM reader.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281473 13f79535-47bb-0310-9956-ffa450edef68