- Added tomcat.core.Container - it's similar with Catalina container, i.e. it represents a group of URLs. It is also similar with per dir config in Apache.
For each URL pattern ( "Location" ) we can associate a handler ( via mappings ) and security
constraints. The Container will represent the properties of a location ( or URL pattern ).

- Added SecurityContraints to Context, use Container to represent them.

- Changed SimpleMapper to also check if the location has secuirty requirements. The code in
SimpleMapper needs a serious review and eventually it should be replaced ( the alghoritm for
matching and the data structures are at least ineficient ).

- Tested 1/2 of security - if you define a constraint in web.xml, it will be processed and
all requests that match will be "marked". The missing part is authenticating and checking if
the user is in one of the authorized roles ( this is done in Craig's code, we just need to merge)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281525 13f79535-47bb-0310-9956-ffa450edef68