More security. Don't expect anything untill 3.1, the code is probably buggy and incomplete.
Basic Auth works ( i.e. the protocol part - no real authorization ).

To realy authorize you can extend SecurityCheck and override "checkPassword" and "userInRole"
or implement a new interceptor based on SecurityCheck that will do whatever auth it needs to.

Tomcat will impose no restrictions / APIs on how the authentication works, you can implement
whatever scheme you want as an interceptor ( and JAAS or any other API for password and role
checks).
There are 2 callbacks you need to implement:

- authenticate : if you recognize a request and feel you can authenticate,  do it and set the
"remoteUser" field. ( XXX if the password is wrong return error - not implemented )

- authorize: the request will have some "roles" set in, and the user. You just need to check
if the user has one of the authorized roles. The same method will be called to implement the
isUserInRole() API call ( with internal request with User and Roles fields set, but nothing
else).
If the user is not authorized you need to set the handler to a servlet that can perform
authorization ( either the internal AuthServlet that will do Basic, or a form, or something
else).
( XXX that can move up into ContextManager - it's common enough )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281526 13f79535-47bb-0310-9956-ffa450edef68