Moved work-dir initialization in DefaultCMSetter. WorkDirInterceptor will only clean up the workdir if not persistent.
Commented out the workdir removal - it's not part of the spec and it affects the
performance ( by removing the jsp files). It is also a bad idea to clean it up
all the time we start/shut down - it can be removed when we un-deploy or
when we have a good reason for that, it is the only place where servlets can save files.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281527 13f79535-47bb-0310-9956-ffa450edef68