Changed ContextInterceptor and mapper to use a simpler and more extensible model.
A Container represents the common properties of a set of URL paths, and we use
this representation for both servlet-mappings ( urlPath -> handler) and
for security properties. The model is extensible - it is possible to add
more complex patterns ( including matching virtual hosts, source IPs, etc),
and it is possible to fine tune the whole server.

This is also similar with Catalina - and will help us share more code.

Also, removed some methods from Interceptor - it's better to have a smaller set
and add later if we need to.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281529 13f79535-47bb-0310-9956-ffa450edef68