Few more errors in error processing.
When we merged the error processing in ContextManager we removed few
workarounds. For example Request should read the headers if
the protocol is >=1.0, even if it doesn't like the URI.
It worked before because it used a sendError() which didn't had any effect
till after reading the headers. Now sendError() has effect, and we
have to use setStatus() and read the headers.

Also, we do print stack traces for all internal errors - that is a very
important tool in debug, and we shouldn't hide them ! Even if it doesn't
look good, if we have a NullPointer we should fix it, not hide it.

As a result you'll see many traces when you run tomcat-test ( problems
with jsp syntax).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281547 13f79535-47bb-0310-9956-ffa450edef68