Fixed the logger - we now use it in Context too, and set it up corectly.
Changed logger names to small caps, no timestams.

Logs from servlet ( via ServletContext.log ) are in a different logger than
internal logs ( you'll no longer see "servet init" messages.

Still a lot of stack traces - will go to a file and it's _important_
to see them, it seems we had a lot of hidden problems.

It seems most of tomcat tests fail because of GTest and the new error system.
The error is reported sooner, and it seems a "Connection reset" is sent after
that. I don't think it's a bug in tomcat, and I'm still looking for a fix in
GTest.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281549 13f79535-47bb-0310-9956-ffa450edef68