- Moved the thread pooling endpoint in PoolTcpEndpoint, the original in SimpleTcpEndpoint. TcpEndpoint is now an abstract class, you need to use one subclass.
- made SimpleTcpEndpoint the default, there are some problems with Pool on
JDK1.2/Linux and it's safer to use the simpler ( Pool is much faster and should
be used in production mode )

- Removed HttpAdapter, use the TcpEndpoint with http adapter. ( HttpAdapter
was a shortcut for that, we want to be able to fine tune http endpoint
and connector )

- Logger - added a "customOutput" property, don't try to do any formatting
if it is set ( no timestamp, no prefix or suffix ).

- Changed "visible" log messages to use the xml style. This will allow
easy post-processing of logs, nice HTML display, filtering, etc.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281553 13f79535-47bb-0310-9956-ffa450edef68