Changed XmlMapper - it no longer ignores errors.
If any error happens in configuration it will report the line number
( if Locator is set and working ) and the xml path / attributes for the
ofending element, and stop parsing.

The caller should handle any recovery or report the error.

This is the same behavior as in the case of a XML error or
DTD violation.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281554 13f79535-47bb-0310-9956-ffa450edef68