Added extra check for getResource() - you can use it to access only resources in the local context. That prevents .. abuse.
Note that a bad servlet can still use File access - this extra check might
catch some bugs and helps if you avoid using File ( which is a good idea for
portability ). In general - getRealPath and File is a bad idea, use it only
when you know what you're doing.


This check can be removed after a SecurityManager is installed.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281625 13f79535-47bb-0310-9956-ffa450edef68