Warning: change to javax.servlet.http.HttpServlet -- all eyes should review.
The If-Modified-Since logic in HttpServlet didn't always behave as it should.
For one thing, it didn't round the millis value down to the nearest second.
This change fixes that problem.  The old code also called
sendError(res.SC_NOT_MODIFIED) to send the 304 which doesn't always work since
Not Modified is not an error status code.  Now it calls setStatus().

I also removed the logic that sent the page if the client's If-Modified-Since
header was after the current system time.  Nowhere in the HTTP spec do I see
that behavior endorsed, and I can't think of a good argument for keeping it.

Please check the code to make sure it makes sense to you.  Feel free to run
your own tests too.  A bug here affects all servlet engines!

-jh-


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281645 13f79535-47bb-0310-9956-ffa450edef68