- Fixed bug in ContextManager.createRequest() - the problem was that we assumed the new request will still be in the same context, which broken various methods on the root context or on sub-contexts.
- That will fix the getPathTranslated() to conform with the current
interpretation of the spec - the result is _different_ from what Apache, ISAPI
and NSAPI return for PATH_TRANSLATED. IMHO it's a serious bug in the spec,
but it's up to the experts to decide what's good for users.
Note that as a side effect, getPathTranslated() is no longer a simple
request attribute that you can get, but it creates a sub-request ( that
means the execution time if you call getPathTranslated() is double).
For Tocmat+Apache,IIS,NES it will create an internal request - the result
might be wrong and will be slow ( since we can't use the PATH_TRANSLATED that
is computed by the server).

- fixed 404 bug in Invoker - if a class is not found, don't send the "class not
found" exception to the ErrorServlet. ErrorServlet generates a 500 Internal
error if an exception is found.

- fixed few tests that were broken - we now pass all tomcat and watchdog tests
for servlets.



PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281691 13f79535-47bb-0310-9956-ffa450edef68