Fix for #116 - JNDI ( and probably EJB and other JDK1.2 libraries ) are not working because the thread ClassLoader is not set.
This bug affects only people using JDK1.2, but I think it's important to check
in the fix. We're using reflection to invoke jdk1.2 methods, and it seems to
work fine in JDK1.1.6.

Thanks to  kyle.downey@ny.frb.org - it's a very tricky bug.
( the solution I checked in is based on his patch, but I used reflection
instead of normal method call to keep 1.1.6 compatibility ). We just need to
call Thread.setContextClassLoader() before executing any servlet method.
( we also need to revert to the previous context class loader if we are in
include - but that happens after the request is completed so it doesn't affect
the response time )

We might need a better solution/architecture to deal with that - please -1 the
change if you feel it's too much for Tocmat 3.1.


Submitted by:	 kyle.downey@ny.frb.org


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281696 13f79535-47bb-0310-9956-ffa450edef68