Commented out few tests - all of them hangs the testing if run with apache.
The reason is that Apache assumes HTTP/1.0 on all requests except clear
HTTP/0.9 requests, and it will deal with all "strange" requests as 1.0.
The test suite doesn't send the ending CR unless the request is HTTP/1.0.
Tomcat forgives it, Apache doesn't - but the HTTP is silent on this issue
( it requires the server to recognize valid 0.9 requests).
Probably Apache is right - the request may be an "extension" or later
version ( post 1.1 ), and all post 0.9 protocols are using headers.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281713 13f79535-47bb-0310-9956-ffa450edef68