Fix for Bug 108, submitted by Larry.Isaacs@sas.com.
PageContextImpl.getAbsolutePathRelativeToContext did not consider the fact
that this could be an included request. If this is an included request then
we need to use the request attribute javax.servlet.include.servlet_path to
make a / rooted path. In the past we were just using request.getServletPath
to make a / rooted path.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281747 13f79535-47bb-0310-9956-ffa450edef68