Added more example users to the tomcat-users file to demonstrate the syntax options.
Modified the examples web application so that BASIC authentication is the
default, but left notes in web.xml about how to easily switch to form
based authentication.  No changes to the protected application itself
are required when you switch -- one of the really nice things about
container managed security :-).

Enhanced the page displayed inside the protected area to confirm the
user's remote username and principal username, as well as allowing the
lookup of roles assigned to this user.  (You wouldn't do this in a
real application, but it is quite useful for debugging your
tomcat-users.xml file contents.)

Removed a debugging output statement from SecurityCheck.java.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281755 13f79535-47bb-0310-9956-ffa450edef68