Add initial cracks at two new documents -- nothing here is authoritative until discussed and agreed to by the TOMCAT-DEV community.
"RELEASE-PLAN-3.2" is a prototype document for the release plan (schedule)
for the next version.  It will be renamed RELEASE-PLAN once we are through
with the current one, which is (naturally enough) for 3.1.  You can see
that I have not tried to identify any specific functionality to be
scheduled for particular releases, but I've laid out a proposed schedule
that gets us to final release by JavaOne 2000.

"STATUS.html" is an accumulation of action items that I've heard, or
talked to people about, that might be interested to add to Tomcat in
version 3.2 or some later version.  The idea is that people volunteer
to work on one or more of these items, and the interested parties discuss
how to do it (on TOMCAT-DEV so other folks can see what's being thought of
and discussed), and then perform the changes.

A TOMCAT-DEV announcement about these two files will be sent in a couple
of minutes.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281763 13f79535-47bb-0310-9956-ffa450edef68