These changes implement our choice (as agreed on TOMCAT_DEV on April 7) to temporarily change the authentication method that Tomcat recognizes for form based login from FORM to EXPERIMENTAL_FORM.  This choice was made to make it easy to continue debugging and implementing this feature, but prevent users from assuming that it works correctly just because it partly works.
To reverse this change later, just change EXPERIMENTAL_FORM back to FORM
in the places listed.

NOTE:  This change recognizes a non-standard value in the web.xml file,
and is intended *only* to assist developers in debugging this feature.  It
is not expected to be supported by 2.2-compliant servlet containers.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281768 13f79535-47bb-0310-9956-ffa450edef68