Added the (old, experimental) ajp23 worker.
Ajp23 was used to test bi-directional communication between apache
and tomcat - the main difference from ajp12 is that the response is
sent as a sequence of blocks ( prefixed by length ) as oposed to a stream
terminated by socket close.

That allows the reuse of the connection and bi-directional communication.

Apache will start by sending a MsgBuff containing the request and wait for
a MsgBuff from tomcat. As soon as it receives one it will start processing it
and send back a response ( if needed ), and so on.

It seems jk_ framework is very easy to work with, but I still have to figure
how to keep the session open ( in a multithreaded environment ) - probably
we need a per/thread cache or something.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281802 13f79535-47bb-0310-9956-ffa450edef68