- the code to recycle object and use only one Request/Response/BufferedOS per thread is back in. It was removed because it created some problems ( I think in 3.0 ), now it's time to fix the real problems and remove the workaround.
- request/response initialization is now in contextManager.initRequest().
That method must be called from interceptors to make sure the new objects
are "wired-in".

- Added ServletOutputStreamFacade - it's not yet used, but it contains the
right implementation(s) for print() ( with real encoding ), and it's cleaned
up of any "internal" methods. The Facade is important to isolate tomcat
from the servlets.

- Added ServletWriterFacade - it's a facade for the PrintWriter and will be
used to enhance object reusing and better buffer support. Right now it contains
only some accounting code.

- Added a number of counters in request to allow request timing. There is no
overhead in normal operation - but if you plug in the AccountingInterceptor
you can get a trace of execution times.

- moved servlet-api-specific logic from ResponseImpl to ResponseFacade. That
will allow us to respect the spec restrictions even if a different Response
implementation is used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281815 13f79535-47bb-0310-9956-ffa450edef68