Fix for destroy() - note that even with this fix, reloading is _not_ production ready and is _not_ a "supported" feature of tomcat 3.1.
It is a new piece of code, added recently with little review ( 2-3 people),
and almost no testing besides trivial cases. Do not expect it to work -
it's checked in to enable simple development and get feedback for
next release ( if someone will have the time/wish to complete this feature).
( it's 2-3 hours of hacking, plus the excelent class loader from jserv )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281817 13f79535-47bb-0310-9956-ffa450edef68