- Cleaned-up ( a bit) the InvokerServlet. Removed some duplicated code ( string.length(), lookups for the same element, etc). Most important difference - invoker will notify the context of a new map, so next time the servlet will be called we'll no longer call 2 servlets instead of one. This alters the old logic a bit - you can no longer test if a request is "Invoker-based" by comparing the servlet with invoker servlet. That test was bad anyway - there is no guarantee that a particular invoker will be used.
- Added logic to get the source of a servlet definition - it can be web.xml,
the admin interface, a dynamic mapping like invoker - and hopefully soon
jsp can take advantage of this.

- Added counters to ServletWrapper - to be used in tomcat performance tuning
( the overhead is very small, and the information may be very usefull even
at run-time. The web interface later )

Please test and review the new Invoker.

( during this review it showed up that exact matching is broken, and it also
look like suffix mapping with PathInfo will not work even if you declare the
JSP page.  )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281827 13f79535-47bb-0310-9956-ffa450edef68