- Renamed TomcatHome to InstallDir in ContextManager ( it generates a lot of confusion ). Old methods still exist, but will be removed if nobody is using them.
- moved deprecated methods at the end, preparing for removal.

- Added comments in ContextManager and more comments in package.html about
how it works.


Problem to be resolved:
Port/HostName do not belong here - tomcat can't guess what is the real
name and port for the default server ( unless it's working as a standalone
http server).  The only use for those properties is in generating the
temp. work dir name. I think we should remove them ( the adapter will provide
the real values for request ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281845 13f79535-47bb-0310-9956-ffa450edef68