- removed CookieUtil - not used, replaced by CookieTools.
- Fixed MaxAge to Max-Age, thanks to "Peter Johnson" <peter.johnson@lutris.com>

- Removed some of the useless object creation. We now use static fileds for
everything that is static ( no need to create a get GMT timezone for
every request, no need to create a HttpDate instance which creates a new
Calendar which create a new Date, use StringBuffer instead of String, etc).

Right now the only objects that are created are a StringBuffer and a Date.
It is possible to avoid both by keeping a per/Request instance that will
be reused ( since all processing happens in the request thread - at least what
 is happening in tomcat, the servlet may create threads but that doesn't affect
us).

- added few more comments

Submitted by:   "Peter Johnson" <peter.johnson@lutris.com>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281868 13f79535-47bb-0310-9956-ffa450edef68