- Removed ServerSession - it wasn't used in a while, StandardSession has more functionality.
- Removed dependencies on tomcat/catalina, removed tomcat/catalina. The package was an
early attempt to keep the 2 projects in sync - now it's outdated and too confusing.

Right now the session management in tomcat is implemented in StandardSession and
StandardManager. The link between tomcat and those 2 objects is SessionManager.

SessionManager will become a per/Context object ( as requested and agreed few months
ago ).

- Reaper still exist in util - it's usefull in itself, will me made a general util.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281896 13f79535-47bb-0310-9956-ffa450edef68