- Removed FixHeaders Interceptor - it was a hack that ended up with duplicated and inefficient code.
- All Content-Type, Length, Language are set in ResponseImpl, the code was redundant.

- Cookies - moved addHeader ResponseImpl too, consistent with the rest of the code
( we add the header for setContentLength, we should do it for Cookies)

- Date - moved the code in HttpResponse, Apache will set the date anyway.
XXX check if NES, IIS set the date header automatically, change the C code to
set it ( date formating in java is very inefficient, and the server is optimized
anyway for the format used in HTTP ).

- Servlet-Engine header will be set in the adapter, before sending the headers.
This allows per/adapter customization ( Tomcat/JNI/Apache ), and we can move it
in the server side, instead of sending it per/request.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281930 13f79535-47bb-0310-9956-ffa450edef68