Added the security patch from Glenn Nielsen (glenn@more.net).
I did few changes and made it work with jdk1.1.8 - of course security will not
be available, but you can still use all other features.

A new server2.xml file will be used for all the JDK1.2-specific components. The old server.xml
will keep all stable and JDK1.1.8 components. IF you use JDK1.1 you'll need to
specifiy -f server.xml flag.

You should be able now to run "untrusted" code in tomcat. Please note that this is NOT
a supported feature - security is a long process and needs a LOT of review. There are still
few components that have no facade ( class loader, the I/O streams), that will
be fixed soon, and also core will be made more restrictive to prevent access.

There are many things to do in this are - but most important is to read the code and
understand how it works, and make sure we're safer than a browser running applets.

Submitted by:	Glenn Nielsen (glenn@more.net)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@281944 13f79535-47bb-0310-9956-ffa450edef68