here is some of the email discussion that should explain these fixes...note that catalina will also probably need to be fixed. craig, i will leave that one up to you. :-)
FAT kudo's to Sean for tracking this one down...it wasn't easy.

-jon

On Wed, Jun 14, 2000 at 07:12:26PM +0000, Sean Legassick wrote:
> The sessions aren't being nullified - they are still there. At least
> part of the problem is that SessionSerializer.doSerialization doesn't do
> anything with sessions when it has been serialized & deserialized, so
> the old hashtable remains in place.
>
> This works for anything loaded from the system classpath, but anything
> loaded by the ACL gets a ClassCastException when it is retrieved from
> the session, as the class of the object retrieved belongs to the old
> classloader. This happens in RunData.getUserFromSession for us, so we
> just start with a new user object.
>
> I'm not sure this is the whole problem

Hi Jon, I thought I'd carry on ....

It's not the whole problem, although it's most of it.

Even when I patched SessionSerializer to replace the old sessions
hashtable in sessionM with the new one, I was getting a
ClassCastException in RunData.populate.

The reason is that when the session is serialized/deserialized the
Request object has already been setup with the current session (with
objects loaded by the old classloader). What's needed is to have the
reload handler also replace the session in the current request.

I've done a quick'n'dirty patch that does this, and it now works for me!
Basically my patch extends ContextInterceptor.reload to have 'Request
req' as a paramter as well, and then I keep on passing the Request down
until it hits SessionSerializer, which can then replace the old session
with the regenerated one there and then.

I'm not sure a Request is the right thing to pass around here, but if
you'd like my patch I can send it to you.

--
Sean Legassick
sean@somacity.com


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282014 13f79535-47bb-0310-9956-ffa450edef68