Fix a "double whammy" bug in Tomcat and Jasper.
The RequestDispatcher.include() implementation was passing an incorrect
value for the "javax.servlet.include.request_uri" request attribute.  It
was only passing the context-relative part of the URI, instead of
including the context path on the front.  This only mattered in non-root
contexts, but will mess up anyone relying on the correct value there.

Jasper had a related bug that was hidden by this Tomcat problem.  If you
include a JSP page with a reuest dispatcher, it was using the value
retrieved from "javax.servlet.include.request_uri" to grab the path to the
page, instead of the "javax.servlet.include.servlet_path" attribute.
Fixing the Tomcat bug described above caused this to fail -- the net
effect was that it doubled the context path in trying to calculate the
real path to the JSP file, which caused FileNotFoundExceptions.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282031 13f79535-47bb-0310-9956-ffa450edef68