Finally a rewrite of DefaultServlet... It had an amazing number of duplicated checks, a lot of complexity.
The new code is a normal interceptor and 2 very simple ServletWrappers, one for
static files and one for dir lists. The default handler needs informations like
the welcome files to operate - so it can't be portable. It is also very inefficient
to have too many layers - the new interceptor just uses the internal API,
without things like try{ res.getWriter() } catch() { res.getOutputStream() }

As a side effect, a number of small fixes, some security checks from DefaultServlet
are now integrated in getRealPath() checks ( FileUtil.safePath() ).

Also, added code to track the include depth ( with hope that a cleaned  RequestDispatcher
 will be ready before code freeze )

Note that Context.getDefaultServlet() no longer exists ( there is no such requirement in the
spec, and that makes standalone more consistent with integrated ), and we use "null" as
the wrapper for a request that can't be mapped.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282042 13f79535-47bb-0310-9956-ffa450edef68