JspInterceptor - experimental code, it will not be enabled by default.
It try to replace JspSerlvet/JspLoader, with an optimized critical path and
simplified operation and a lot of caching.

It works in a similar way with JspLoader:
- when a change is detected, a new class is generated (instead of a classical
reload ) - all classes have a version number
- after restart, we find what was the latest version and use it ( if no change,
we'll not recompile )

TODO:
- implement precompile
- implement destroy ( before compiling the new servlet, destroy previous one )
- name mangling - we use a simple scheme ( i.e. jsp dir -> package, jsp file name
+ version -> classname ). Probably we don't need to generate the same mangled
name, but we should add the code to avoid special cases like java keywords or invalid
names. IMHO the simple name is worth keeping when no special case is detected

Note that this version should corectly reload if a class used by a jsp is changed.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282085 13f79535-47bb-0310-9956-ffa450edef68