- Added ReloadInterceptor. As promised, the reloading will be re-written and improved, and this is the first step. Right now we implement almost the same scheme as before ( so it's not a fix ), but we can move the whole mess in a repaceable component.
- Added DependManager. It is a (stand-alone) component, with no deps on
tomcat that is able to detect a change in a set of files. It will be
improved in time, but so far it provides enough for tomcat.
It is able to minimize the number of file accesses and perform good
enough.

- Context: on reload, all servlets will be cleaned up.

- started to phase out ServletLoader. AdaptiveServletLoader will still
be around ( to ensure continuity and stability ). The interface will be
removed in few days unless someone is using it ( for new class loaders)

- removed deprecated methods in context.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282225 13f79535-47bb-0310-9956-ffa450edef68