ByteBuffer was an experiment to improve in/out systems in tomcat. It is not used in tomcat3.2, but can be enabled and shows some decent improvements.
On the other side I think it is too dificult to use, and I will roll it back until I find
a better solution.

I did a re-write of BufferedServletOutputStream, keeping only the buffering
code in OutputBuffer and moving all tomcat-state management code in ContextManager
and Response. This will allow alternative implementations of OutputStreams ( so
far BSOS was the only one ).

BufferedServletOutputStream is no longer used - it will stay around until we make sure there are
no people needing it.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282240 13f79535-47bb-0310-9956-ffa450edef68