Bugfix: a packaged tag library can not be specified in web.xml.
This was not consistent with JSP 1.1 spec errata issues 8 and 9. Now the algorithm is closer to what is suggested in the errata:
First the TLD_URI is resolved and then it is checked whether this URI represents a .tld file or a .jar file. In case of a jar file, META-INF/taglib.tld entry is used.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282241 13f79535-47bb-0310-9956-ffa450edef68