Clean up the former special handling of getContext("/") calls.  Now, requests for this are treated like anything else (and will return the root context).
Add a new Context property (crossContext) that allows the administrator
to turn on and off, on a per-app basis, the ability for getContext() to
cross application boundaries.  Defaults to "true" to keep Watchdog happy.

The Tomcat 4.0 version of Catalina passes all the Watchdog servlet tests
except for the getContext() one, but that is only because the ROOT context
is not currently configured in $CATALINA_HOME/webapps.  That will be fixed
soon.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284961 13f79535-47bb-0310-9956-ffa450edef68