- removed dead code, small cosmetic changes ( incl. few more comments).
- removed Constants that are no longer used.
 Moved ATTRIB constants in the Context ( where get is called ) - first to insure
consistency and also to be able to document them ( comments ).

- same for Response defaults, session defaults.

Also planed: move ServerName/version to ContextManager ( wich is the actual
server ), and remove Constants. All constants should be defined in the component
where they belong ( request attributes in request, context attributes in context,
etc). This allows us to keep the code in sync in the future.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282266 13f79535-47bb-0310-9956-ffa450edef68