Important fix - transport constraints checking. This is a spec requirement, I think it should go in.
As allways - please let me know if you have any doubt and I'll roll back.

In AccessInterceptor we'll check if CONFIDENTIAL/INTEGRAL is required
for a request and if the SSL is not used ( isSecure() is false) we'll
not serve the request ( using 403 - Forbidden ).

The 403 handler is able to do a redirect to the SSL server. The secure
port must be configured using ContextManager.setSecurePort().


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282280 13f79535-47bb-0310-9956-ffa450edef68