Correct the startup call order to ensure that the application class loader is completely initialized prior to being required to load the application event listener classes.
Add example context and session listeners to the "/examples" webapp.  They
simply log the occurrence of interesting events using ServletContext.log.

WARNING:  There is a bug in the current version of the 2.3 servletapi
classes that causes HttpSessionBindingEvent.getSession() to return null.
This causes NPE exceptions in SessionListener, but the application
continues to operate so long as it does not rely on getSession().


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284999 13f79535-47bb-0310-9956-ffa450edef68