Implement substantial changes to ApplicationDispatcher (the request dispatcher implementation class for Catalina) to support the new 2.3 ability for servlets to wrap the request and response objects that they pass to forward() or include().  Previously, servlets were required to use the original request and response objects.  This is the largest single impact change required by Servlet 2.3.
Be sure you pick up the updates to the "jakarta-servletapi" module, which
fixes a problem with getSession() returning null on all session related
events.

As of now, Tomcat 4.0 passes all but one of the Watchdog servlet tests
(the one that checks for version 2.2 :-), and all but one of the JSP tests
(positiveInclude.jsp, which needs further investigation).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285001 13f79535-47bb-0310-9956-ffa450edef68