More hacking on build.xml.
Now it is possible to compile tomcat with either servlet23.jar or
servlet22.jar. It will work only with servlet22.jar, of course.

Also checked in the old facade ( with the required aditions )
into facade23, to get it started. I hope all facade23 code will be
re-written with a fresh implementation.

In order to have tomcat working with 2.3 the facade23 must be
implemented and few specific interceptors, plus a modified
server.xml ( that will use the 2.3 interceptors, and set the 2.3
facade into all contexts - XXX make the default configurable ).

Running both in the same VM is still a goal, but it's not the
biggest priority. Separating the facade from the core is more
important.

The old facade code and 2.2 specific interceptors
should move to facade22 directory.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282310 13f79535-47bb-0310-9956-ffa450edef68