Original rationalle for this change captured for posterity:
In order to support JSPs (not precompiled) in our embedded environment we need
to get the jar files associated with tomcat (webserver.jar, etc) and our
embedding application jar files (APIs needed) into the path searched by the
compiler. We have added a system property (tc_path_add) with these jar files
into the list passed to the compiler in Compiler.java.  We don't want to modify
java.class.path because other 3rd party code also running in the JVM.

These jar files are not in the java.class.path system property since we use our
own classloaders (URLClassLoaders). Based on a comment in the code there seems
to have been a jsp.class.path system property in the past that was removed that
may have been used for a similar purpose. It is possible to get this added back
in or is there an alternative solution in the works for this type of problem in
the 3.2 timeframe?

Submitted by: Mac Holloway <Mac_Holloway@tivoli.com>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282361 13f79535-47bb-0310-9956-ffa450edef68