Step 2 in MimeHeaders refactoring.
MessageBytes is now exposed and used instead of Strings. Most methods
returning string are removed - if we do the conversion too early (before
charset is detected ) we loose all the benefits and optimizations of
MessageBytes.

As proposed earlier, MessageBytes will gradually replace all Strings
- toString() and setString() can be used to provide the old functionality.

Also, rewrote the code to enumerate the header names and multi-values.
The new code doesn't create GC ( Hashtable, Vector ) and it's probably
(more) correct.

I think the new code is manageable now, looking forward to review and
feedback.

( BTW, the new methods are now getName() and getValue() instead of
getHeaderName, etc. The reason is that the same object can be used
to store the parameters in an efficient way)

This shoud give a significant performance improvment when completed, but the
main reasons for doing those changes are:
-support for non-ascii charsets
-refactoring - better, readable code


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282368 13f79535-47bb-0310-9956-ffa450edef68