Added few test files I had in various places on my machine. I used them to verify various things, but never had a chance to write a formal <test> - you may find them useful ( or ignore them if not) ( some are based on bug reports - sorry if I missed the contributor )
- Hello, HelloW are good for testing server overhead ( simple pages
with minimal time spent in the servlet, all the time processing
the request)

- sessionShare used to verify that sessions created in servlets are
visible in jsps

- Policy tests tomcat.policy for servlets ( I also use to build a jar and
put it in lib to verify the policy applies to classes and lib too )

- XXXStream, Writer are usefull to evaluate the overhead of the 3
output methods ( ServletOutputStream/bytes, SOS/chars, writer). Very
interesting ( or it used to be in 3.1, 3.2 ).

- RmiTest - can you start RMI from a servlet ?

- CookieExpire - do we set the cookie expire ?

- Post - I'm not sure if it is supposed to work ( it hangs until you hit
stop in the browser ).

Again - if you don't like them feel free to remove, I find the tests
usefull even if the code is not very clean - just hacks.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282379 13f79535-47bb-0310-9956-ffa450edef68