Work around a problem with Java clients using the JDK 1.3 version of the HttpURLConnection class (and probably other HTTP/1.1 clients as well).
The problem was that the client would send an HTTP/1.1 request, and include
the "Connection: keep-alive" header.  Catalina's HTTP/1.1 connector would
therefore dutifully keep the connection open and listen for a subsequent
request, which would not happen if the client just closed the connection.
Now, the request processing code detects this situation and silently closes
the socket, instead of complaining about the lack of a request line.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285031 13f79535-47bb-0310-9956-ffa450edef68