Do not include the JSSE jar files in $CATALINA_HOME/server -- due to the way that class loading is implemented in Catalina, JSSE *must* be installed as a system extension (in $JAVA_HOME/jre/lib/ext), as described in the JSSE documentation.
You can now enable simultaneous non-SSL and SSL connectors with Tomcat 4.0
stand-alone, once you follow the directions in the "conf/server.xml" file
and uncomment the entry that enables this connector.  The default SSL port
is 8443 -- you won't be able to run on port 443 on Unix systems unless you
are running as root, which is not a good idea.

Next step is to add initial logic that deals with client authentication
and certificates.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285050 13f79535-47bb-0310-9956-ffa450edef68