Changed the location of jk and jserv files - now it's conf/jk, conf/jserv.
I don't know if this is the best location, but it's a step forward ( IMHO ).
( the pattern is similar with /etc, /etc/mail, /etc/www, etc organization -
each "component" have a directory ).

( I am happy with any conf/ organization that groups the files somehow -
please change it or -1 if you want )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282407 13f79535-47bb-0310-9956-ffa450edef68