Remove URL decoding of the entire request URI.  This conforms with the required RFCs, as well as the "errata" to the 2.2 servlet spec that says getRequestURI() should return a value that has *not* been decoded.
Currently, Tomcat 4.0 does not decode any of the path components -- there
is an ongoing discussion in the servlet experts group to clarify what the
final rules will be regarding what gets decoded and when (i.e. before you
do the mappings or after).  Once final resolution of these issues is
reached, any needed changes for the required behavior will be done.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285071 13f79535-47bb-0310-9956-ffa450edef68