Add showStackTraces attribute to the ContextManager to control the display of stack traces in the default response for exceptions.
Added showStackTraces="true" to ContextManager in server.xml to
help show that this feature is present and on by default.

Added a space after the colons in the resource text strings used in
the exception response.  When stack traces are disabled, the
exception message will follow the colon on the same line.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282449 13f79535-47bb-0310-9956-ffa450edef68