CHANGE TO THE SERVLET API!
Per the Servlet 2.2 errata published at
http://java.sun.com/products/servlet/errata.html
the status code 307 (SC_TEMPORARY_REDIRECT) was
supposed to be added to HttpServletResponse, but
it never was.  I just added it.

Please do a double check that all's OK with the
change.  This is something that should go into
the 3.2 release if possible.

I'll email Danny/Eduardo to see if they want to
push another servlet.jar as well.

-jh-


git-svn-id: https://svn.apache.org/repos/asf/tomcat@266766 13f79535-47bb-0310-9956-ffa450edef68