Start session refactoring.
ServerSession was part of the original Tomcat3.0 design, I replaced it with
StandardSession from Catalina. That created a number of problems, and it's
time to solve that.

We now use the same patterns for session as for the rest of tomcat - 2
layers, the lower layer uses internal objects ( for speed, recycleable,
focused to the core functionality ) and the facade aggregates core
objects and implement servlet-specific behavior.

Note that session is broken today ( I'll finish it this evening !). I try to
get "major" changes  done ASAP in order to be able to have a freeze and
spend the beta time fixing all bugs.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282463 13f79535-47bb-0310-9956-ffa450edef68