Finally, src/share is completely independent of servlet.jar  and we have a clear layer and module separation.
Tomcat.util is the collection of general-purpose, independent utilities

Tomcat.core represent the web layer ( low-level objects, close to web
server architecture ) and depends _only_ on utils and JDK1.1

Tomcat.modules consists of general modules (independent of servlet.jar,
but acting as adapters for external APIs and tools ), depend on core,
utils and the APIs they plug in ( some may be JDK1.2-specific )

Tomcat.facade is the adapter between core and a certain ServletAPI. Since
core maps ( almost completely ) to all common web servers, and servlet API is
supposed to work with those servers - all servlet API versions can be
supported on this architecture ( including legacy 2.0, and for sure 2.3
and whatever follows - including 3.0 ).

There is only one extension mechanism ( BasicInterceptor ), that mapps to
the extension mechanisms used by common servers, and about 5 core classes that
matter.

Interceptors are the basis of tomcat modularity, and we can tune up individual
components.


:-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282467 13f79535-47bb-0310-9956-ffa450edef68