Ops, forgot that one too :-)
- fix facade23 ( it compiles and should work 2.2 apps, the new 2.3 features
need to be implemented - again, I'll try to stay away from this task
as much as possible, it's better to have a fresh look at facade refactoring )

Added a hack to allow compilation of both facades - checked in binaries.
I know it's bad, but it's very hard to build tomcat when it requires 2
different cvs branches that happen to build in the same dir.

I'll try to make changes so that only servlet2.2 needs to be checked in (
probably it'll not change - it's closed now ), and we can build 2.3
, but right now I prefere a simpler solution.

( of course, I'll have to remove them if you -1 it ... )

( use tomcat-new target to build both !!!! )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282470 13f79535-47bb-0310-9956-ffa450edef68