Fix OutputBuffer bug ( at least one :-)
The problem was simple - recycle() didn't reset the buffer of the
writer ( convertor ), and in some cases ( for example if the connection
is broken ) the buffer was not flushed.

XXX Need to find out why the buffer wasn't flushed - I think we
should assert( ob.flush() called after each request ).

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282471 13f79535-47bb-0310-9956-ffa450edef68