Removed FacadeManager.
FacadeManager was used to allow the separation of core and facade, and
also as a mechanism to access the real Request from a facade.

The first use is replaed by normal interceptors - there is no need to use
a special mechanism and interceptors provide much more control and power.

For the second use we don't actually need the FacadeManager - instead of
calling getAttribute() to get FM and then getRealRequest(), we can just
get the Request directly from getAttribute() ( and same checking are done)

The change also moves WebXmlReader to facade ( since it's specific to
servlet2.2 and 2.3 will have a different version )
We add the Servlet22Interceptor - still need to add code to support
multiple facades at the same time.

During this change I also fixed a bug in JspInterceptor and added a number
of debug statements.

Another related change ( needed to support recycle() of facades ) is
the postRequest() callback. postService is called after all servlets, including
included - postRequest is a better place to log and clean up after
a client transaction ends.

Expect a bit of instability after this change - will be fixed during next
few days ( and weekend ) . The code is still the same :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282479 13f79535-47bb-0310-9956-ffa450edef68