Almost done with Context( first out of 8).
Moved getRealPath() and getResource() to facade - both are likely to
be specific to version and not used in core.

GetRealPath() was called in StaticInterceptor and few places -
it all ends up to safePath(), a very expensive operation - and it's
bad to call it several times instead of once. Need to
optimize SI. Anyway, it's better to have the bad code clearly
visible.

Also added more documentation to Context, re-aranged the code, added
short comments to fields.

Except for addContainer() and addServletMapping() everything is done
( from my point of view - I'm waiting for the next who likes some fun
refactoring code - I hope there isn't too much left in Context :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282488 13f79535-47bb-0310-9956-ffa450edef68