Implement a new object hierarchy level in server.xml called <Service>.
A Service is defined as a set of one or more Connectors that share the
same Container (in most cases an Engine) to process their requests.  For
example, Tomcat 4.0 running stand-alone might typically have a non-SSL
connector running on port 8080, and an SSL connector running on port 8443.
At the same time, a separate Service might be present with an Apache
Connector talking to a specialized Engine imlementation that knows how to
share configuration information back to Apache (this connector is nearing
completion).

In addition, a new classloader level for shared server extensions (i.e.
the contents of the $CATALINA_HOME/lib directory) is now propogated down
to the Context implementations (via setParentClassLoader) so that it
becomes the parent classloader for all webapp classloaders.

Remaining steps (later today):

* Add a developer document describing the class loader architecture
  inside Tomcat 4.0.

* Enforce the checking for required packages and versions as described
  in the new spec (the infrastructure is all there now, just not the
  enforcement).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285143 13f79535-47bb-0310-9956-ffa450edef68