Second patch from Stef�n Freyr Stef�nsson <stebbi@decode.is>.
Make javax.servlet.request.X509Certificate available - this is part
of the spec.

In order to do that, 2 new methods are added - that can't have any negative
impact or add regressions. First ( getServerSocketFactory in Endpoint )
allow clients to directly access the factory used by the endpoint.

Second, preProcessRequest() allows a factory to set attributes and
alter the request before it gets processed.

In the case of non-ssl http nothing change - just an empty function call.
If SSLSocketFactory is used, it'll just add the attribute ( and also
set scheme to "https" - in case you forgot to add "secure=true" attribute
in the config - it's a cleaner way to do that )

The only negative side of the patch is that it adds a dependency between
SSLSocketFactory and tomcat.core.Request. I think this is not a problem -
tomcat 3.2 doesn't claim or tries to support this as an API, and
we know that in 3.3 everything will be clean ( by using ssl interceptor
to set the attributes )

Please let me know if you see any problem with that, and I'll roll back.
I don't think it'll affect the current functionality in any way.

Submitted by: Stef�n Freyr Stef�nsson <stebbi@decode.is>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282502 13f79535-47bb-0310-9956-ffa450edef68