Patch from Sebastien Alborini <sebastien.alborini@m4x.org>.
It has 3 components:

- added a field and get/setParentClassLoader() in ContextManager.
That can't have any side effects.

- In LoaderInterceptor, check if parentClassLoader is set, and if
it is use it instead of getClass().getLoader(). In normal use ( if
cm.parentClassLoader is not set ) it doesn't change anything in the
current functionality.

- In Jdk12Interceptor, nothing change unless "useThreadLoader" is
set ( probably a better name would be useful ). If this property
is set, Jdk12Interceptor will set the ContextManager parent loader
to be the context class loader. This doesn't change anything in
normal operation. The default is false, to match current functionality.

Please review and if you feel that have any side effects I can roll
back.

Submitted by: Sebastien Alborini <sebastien.alborini@m4x.org>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282503 13f79535-47bb-0310-9956-ffa450edef68