Add a new extension of Container called Deployer, which represents a Container into which web applications can be deployed and undeployed.  In the current environment, StandardHost is the only container implementation that is also a deployer -- but additional such containers can be created (such as the forthcoming specialized container for the Apache connector).
Implemented the deploy() and undeploy() functionality in StandardHost.
Now, Tomcat 4.0 is capable of "hot" application loading and unloading,
given just a context path and the URL of a WAR file.  This will be of
tremendous use in administrative and management applications.

As before, application deployment from a WAR file in the application base
directory ($CATALINA_HOME/webapps for the default configuration) will
cause the WAR to be automatically expanded.  However, the new undeploy
functionality also deletes the directory when the application is
undeployed -- the theory is that the developer will be replacing the WAR
anyway in this scenario, and this is a common complaint about the way that
Tomcat 3.x works (once the directory is expanded the user has to remember
to delete it).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285158 13f79535-47bb-0310-9956-ffa450edef68