- Moved doWrite from ContextManager to OutputBuffer ( context manager was an un-necesary middle man ).
- Fix in MessageBytes - reset the cached values when set is called

- First major change in Request - method, uri, query and protocol are
now MessageButes, as proposed some time ago.
This follows the 0 GC goal, and is essential for real encoding and
charset support.

More to come this weekend.

IT seems that most of tomcat 3.3 works again ( more or less - we had some
big changes ). This weekend I'll try to finish request and everything else
in core.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282514 13f79535-47bb-0310-9956-ffa450edef68