Didn't interpret Pierre Delisle's patch correctly.  I was up too late the night before.  Now it's correct.
For reference, here is Pierre's commit message which explains the original problem:

It is the tag <uri> in the TLD that sets the value of the
'urn' field, not <urn>.

And for more details:

Referring to bug 214, Hans Bergsten wrote:
> Looking at the code, it seems like it's a bug that an element named
> "urn" is processed instead of the "uri" element. But I assume this
> only means that the "uri" element is (incorrectly) ignored but that
> the TLD is still parseable. Since the "uri" element is intended to
> be used by development tools rather than the container, this should
> not be a fatal problem.

Correct.

I must admit the spec is quite confusing on that matter.

The confusion comes from the fact that the <uri> subelement of <taglib>
in the TLD is mapped to the 'urn' field of TagLibraryInfo
(which is accessible via the method getReliableURN()).

This 'reliable urn' is not used anywhere else in the source base.
As Hans mentioned, it is "intended to be used by development tools
rather than the container".

The code simply needs to be fixed by accepting the tag <uri> in the TLD
instead of <urn>.

[Wish the tag was named something like <reliableurn> in the
TLD instead of <uri> to remove the confusion.

And, for completeness, the 'uri' field of TagLibraryInfo is set
via the 'uri' attribute of the taglib directive. It is accessible
via the getURI() method of TagLibraryInfo]


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282517 13f79535-47bb-0310-9956-ffa450edef68