Work around a behavior change in the java.net.URLClassLoader implementation in JDK 1.3.0 for Windows.  If you specify a repository URL like this:
	file:C:\Jakarta\build\tomcat-4.0\.\webapps\scarab\WEB-INF\classes/

this repository would get silently ignored (even though JDK 1.2.2 was able
to use it just fine).  This kind of thing happened when you started Tomcat
4.0 without setting the CATALINA_HOME environment variable.

Now, Tomcat 4.0 parses out the "\.\" occurrence, and Scarab can be started
successfully (which was previously failing).

If you want to run Tomcat 4.0 under Windows JDK 1.3.0 in the mean time, be
sure you set the CATALINA_HOME variable to point at the top-level
directory of the Tomcat 4.0 install for your system.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285188 13f79535-47bb-0310-9956-ffa450edef68