Add support for searching for Localized versions of static files.
In Context, add 'String getRealPath(String, Locale, Locale)' method
to allow to look for a locale-specific real path.

In StaticInterceptor, add constants (static final String) to identify the
localization property name and the valid values.
Define 'void setLocalization (String)' method to allow setting the
localization based on the "localization" property set in the server.xml
entry for the StaticInterceptor.
Define 'String getLocalization()' to the the value selected.
Define 'String getWelcomeFile (Context, String, Locale, Locale)' method
to allow for locale-sensitive search of welcome files.
Modify 'requestMap' to perform localized resource lookup if configured
with the Interceptor, behave as before if not.
Also, modify the logic of getting the welcome files, to allow to invoke
the locale-sensitive version of 'getWelcomeFile'.

In FileUtil, add methods to support locale-sensitive lookup of
resources. The algorithm used is similar to the one specified in
java.util.ResourceBundle.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282534 13f79535-47bb-0310-9956-ffa450edef68