Work around the lack of a JSSE call to ask for the key size.
The cipher suite name is examined to detect cases where it contains a cipher
name for which the key size is known.  If it is known, the corresponding key
size is published.  Otherwise, a key size of zero is published.

The underlying data for this mapping comes from the TLS Version 1.0
Specification (RFC 2246), Appendix C.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285220 13f79535-47bb-0310-9956-ffa450edef68