Refactor the startup processing so that saving sessions across automatic context reloads, and normal shutdown/startup of the server.  Previously there were cases where you would not see the session information restored correctly.
NOTE:  There are still issues with the timing of when application event
listeners are initialized versus when session restoration occurs.  These
will be addressed shortly.

NOTE:  If you have saved session information from a previous version of
Tomcat 4.0, you will not be able to reload it with this version - the
format of serialized sessions has changed slightly.  To avoid any
problems, simply delete the $CATALINA_HOME/work directory before starting
this version of Tomcat 4.0.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285221 13f79535-47bb-0310-9956-ffa450edef68