Update to use Tomcat 3.2 server.xml as the template.
Now assumes Tomcat is unzipped to the root of C drive,
i.e. c:\jakarta-tomcat.  Also, the jni_connect.dll is assumed to be in the
win32\i386 directory under Tomcat's bin directory.  This matches where it is
found relative to the directory where the Tomcat binary distribution is found.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282549 13f79535-47bb-0310-9956-ffa450edef68