Fix a nasty security hole in Tomcat 4.0 that would allow a URL like:
  http://localhost:8080/examples/jsp/snp/snoop.jsp/

or

  http://localhost:8080/examples/jsp/snp/snoop.jsp\

to display the source code of the JSP page.  The same problem occurred
with static files -- the raw HTML code would be returned if there was a
"/" or "\" after the name of a valid static file..

Now, these requests will all return 404 (not found).

Submitted by:	Cheong Takhoe <Takhoe@APIIT.EDU.MY>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285238 13f79535-47bb-0310-9956-ffa450edef68