Enable per-web-app negotiation of the need for client certificates.
Previously, the SSL socket factory had to be configured globally to
require (or not require) client certificates for *all* apps.  Now, you can
leave the factory configured to not require certificates.  Individual
applications that use CLIENT-CERT authentication will automatically
renegotiate the SSL session (with client certificates required).

If all of your applications require client certificates, it is still
faster to configure the factory to require them -- this avoids the need to
renegotiate on every single session to every app.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285288 13f79535-47bb-0310-9956-ffa450edef68