WebDav seems to work fine in DAVExplorer. Excelent code...
The directory structure is a mess, I'll fix it later ( i.e. move
all the sources in src/webdav/WEB-INF/src, make it a standalone
webapplication).

So far it seems to work with absolutely no dependency on tomcat -
if someone has another container it would be nice to test it!!!!
( the .war will be available soon, need to fix the build )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282578 13f79535-47bb-0310-9956-ffa450edef68