Work around a nasty problem that is caused by the way SessionInterceptor is implemented in Tomcat 3.2.
The problem report (BugRat Bug Report #316) indicates that session cookies
were not being set if a servlet did a RequestDispatcher.include() of a JSP
page.

The actual problem would occur anytime that the first flush of the response
buffer occurred inside the included servlet.  It happens because the session
interceptor (that sets the session cookie) uses the same method used by
servlets (response.addHeader()) to add the session ID cookie -- but attempts
to add headers inside an included servlet are ignored.  A JSP page as the
included target triggers this because it includes a flush someplace; you can
easily duplicate it with regular servlets as well.

The workaround is to call response.flushBuffer() before actually calling
the included servlet.  This causes the session ID cookie (and the
Servlet-Engine header) to be correctly set.  It should not cause any
compatibility problems, because included servlets are not able to do anything
but write to the output stream or writer (or flush) anyway.

PR: 316


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282605 13f79535-47bb-0310-9956-ffa450edef68