Restore a limited ability to configure the matching prefix for the invoker servlet (defaults to "/servlet/").  This ability had been removed when Tomcat 3.2 was changed to stop reading the global web.xml file.
NOTE:  The default you configure in server.xml applies to *all* web apps that
are installed in this instance of Tomcat.

NOTE:  Tomcat 4.0 reverts back to the behavior of Tomcat 3.1 -- it uses the
values in "conf/web.xml" as defaults for all web applications, which are then
customized from the web.xml file for each individual web app separately.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282606 13f79535-47bb-0310-9956-ffa450edef68