Restore the previous exception propogation model so that exceptions are propogated only from included servlets.  This undoes part of a change Larry committed earlier -- after further discussion with him, we've agreed that this is the correct behavior so that servlets can trap exceptions thrown by included servlets without corrupting the content of the response.
NOTE:  If you include a JSP page that declares an error page, and your JSP
page throws an exception, the transfer to the error page will still happen
as expected even in an included page.  This is handled completely within
the JSP environment, and did not rely on the error propogation mechanism
for servlet exceptions -- which was Larry's primary concern.

As a result of this change, Tomcat 3.2 will trigger the <error-page>
handling only if the top-level servlet throws an exception.  This is also
consistent with the behavior of Tomcat 4.0 in this respect.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@282626 13f79535-47bb-0310-9956-ffa450edef68